requirements.txt
torch
librosa
ffmpeg-python
pydub
transformers
tqdm
